pyproject.toml
[tool.poetry]
name = "apr2-ts2-grabber"
version = "0.1.0"
description = ""
authors = ["Roman Istomin <ra.istomin@yandex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymodbus = "3.7.4"
pyyaml = "^6.0.2"
datetime = "^5.5"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
